+++ exporter/opentelemetry-exporter-otlp/src/opentelemetry/exporter/otlp/exporter.py
@@ -76,7 +76,7 @@ def _get_resource_data(
 
         collector_resource = Resource()
 
-        for key, value in sdk_resource.labels.items():
+        for key, value in sdk_resource.attributes.items():
 
             try:
                 # pylint: disable=no-member
@@ -186,7 +186,7 @@ class OTLPExporterMixin(ABC):
                 if error.code() == StatusCode.OK:
                     return self._result.SUCCESS
 
-                return self.result.FAILURE
+                return self._result.FAILURE
 
         return self._result.FAILURE